Java HTTP server
A simple Java http server. Run and try localhost:3333/sleepyok in browser / curl

My plan was to connect this to a DB to form the beginnings of a backend. The reason it's in Java is because we have Postgres code from first year and I want to get something working lol