Setup for backend user authentication
Some class shells and a db table for authenticating a user client. We can use this to show off a login feature prototype. I'm planning on generating a token for the user client when they sign in with their username and password. The client will then provide the token in the header for future requests.