One-time-code-based pairing login system
Backend facility for pairing a client with an account, where the client does not have input methods which make it easy to enter the traditional username and password. An already logged in device, e.g. the app, can initiate the pairing process by sending a POST to /pair, which returns a code which can be entered on the new device. The new device presents this to the backend, which then presents both parties with a second code, for the user to confirm that the code is the same on the displays of both devices. Once this is confirmed on the original device, a token is made available for the new device.

Not thoroughly tested yet, but seems to be working okay.